Added Checking if user is renting owner in deleteRenting

+++ controllers/rentings.js
@@ -2,7 +2,7 @@ const Renting = require('../models/Renting');
 const Provider = require('../models/CarProvider');
 
 //@desc     Get all renting
-//@route    GET /api/v1/Rentings
+//@route    GET /api/v1/rentings
 //@access   Public
 exports.getRentings = async (req,res,next) => {
     let query;
@@ -38,7 +38,7 @@ exports.getRentings = async (req,res,next) => {
 };
 
 //@desc     Get a renting
-//@route    GET /api/v1/Rentings
+//@route    GET /api/v1/rentings
 //@access   Public
 exports.getRenting = async (req,res,next) => {
     try{
@@ -58,7 +58,7 @@ exports.getRenting = async (req,res,next) => {
 };
 
 //@desc     Make a Renting
-//@route    POST /api/v1/hospitals/:carProviderId/Renting
+//@route    POST /api/v1/carproviders/:carProviderId/rentings
 //@access   Private
 exports.addRenting = async (req,res,next) => {
     try {
@@ -73,7 +73,7 @@ exports.addRenting = async (req,res,next) => {
         const existedRenting = await Renting.find({user: req.user.id});
         //renting limit
         if(existedRenting.length >= 3 && req.user.role !== 'admin') {
-            return res.status(400).json({success: false, message: `user ${req.user.id} has already made a rent`});
+            return res.status(400).json({success: false, message: `user ${req.user.id} has already made 3 rents`});
         }
 
         const renting = await Renting.create(req.body);
@@ -86,7 +86,7 @@ exports.addRenting = async (req,res,next) => {
 };
 
 //@desc     Edit Renting
-//@route    PUT /api/v1/Renting/:id
+//@route    PUT /api/v1/rentings/:id
 //@access   Private
 exports.updateRenting = async (req, res, next) => {
     try {
@@ -116,8 +116,15 @@ exports.deleteRenting = async (req,res,next) => {
 
         if(!renting) return res.status(404).json({success: false, message:  `No renting with id of ${req.params.id}`});
         
+        //Make sure user is the renting owner
+        if (renting.user.toString() !== req.user.id && req.user.role !== 'admin') {
+            return res.status(401).json({success: false, message: `User ${req.user.id} is not authorize to delete this renting`});
+        }
+
         await Renting.deleteOne();
+
         res.status(200).json({success: true, data: {}});
+        
     }
     catch(err) {
         console.log(err);
